Extract shared functions/types from Custom scanner state machine
